refactor file names and locations
